move build info to common containers
