+++ src/renderer/containers/build_info.tsx
